bug(flags): required flags with default value still failed validation
When a required flag had a default value, execution still failed, saying
that the flag has not been provided.

This change fixes that behavior by ensuring that default values are
taken under consideration when checking whether a flag is missing or
not.

NOTE: environment variables will still override default values.