fix(execute): use fresh context for post-run hooks
This change ensures that command post-run hooks receive a fresh context
instead of the original context passed to the `Execute` functions.

This is needed since often the context passed to the `Execute` functions
is canceled by the time the post-run hooks are ran - either by the OS
signal, deadlines, or other reasons.